CLDSRV-805: Regroup and dedup versioning Get/Put
Exact same test in 2 different files regrouped.
Using 1 bucket on test suite to reduce gcp quota pressure
